# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateServerRequest:
    properties:
      capacity:
        type: integer
      city:
        type: string
      country:
        type: string
      flag_emoji:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      public_key:
        type: string
      server_address:
        type: string
      server_name:
        type: string
      server_port:
        type: integer
      short_id:
        type: string
      sort_order:
        type: integer
      xui_base_url:
        type: string
      xui_inbound_id:
        type: integer
      xui_password:
        type: string
      xui_username:
        type: string
    type: object
...
  handler.UpdateServerRequest:
    properties:
      capacity:
        type: integer
      city:
        type: string
      country:
        type: string
      flag_emoji:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      public_key:
        type: string
      server_address:
        type: string
      server_name:
        type: string
      server_port:
        type: integer
      short_id:
        type: string
      sort_order:
        type: integer
      xui_base_url:
        type: string
      xui_inbound_id:
        type: integer
      xui_password:
        type: string
      xui_username:
        type: string
    type: object
...
  model.Server:
    properties:
      capacity:
        description: Capacity and health
        type: integer
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      current_load:
        type: integer
      flag_emoji:
        type: string
      id:
        type: string
      is_active:
        description: Status
        type: boolean
      last_check_at:
        type: string
      name:
        type: string
      ping_ms:
        type: integer
      server_address:
        description: Server connection details
        type: string
      server_port:
        type: integer
      sort_order:
        type: integer
      status:
        description: online, offline, unknown
        type: string
      updated_at:
        type: string
    type: object
...
host: api.zaruchevskiy.ru
info:
  contact: {}
  description: |-
    REST API of the ZyVPN Telegram Mini App backend. All
    /api/* routes require X-Telegram-Init-Data; /api/admin/*
    additionally requires admin role in the database.
  title: ZyVPN API
  version: "1.0"
paths:
...
  /api/admin/servers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: List all servers (admin)
      tags:
      - admin
    post:
      consumes:
      - application/json
      parameters:
      - description: server config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Server'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - TelegramInitData: []
      summary: Create server
      tags:
      - admin
  /api/admin/servers/{server_id}:
    delete:
      parameters:
      - description: server uuid
        in: path
        name: server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Delete server
      tags:
      - admin
    get:
      parameters:
      - description: server uuid
        in: path
        name: server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Server'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - TelegramInitData: []
      summary: Get server (admin)
      tags:
      - admin
    put:
      consumes:
      - application/json
      parameters:
      - description: server uuid
        in: path
        name: server_id
        required: true
        type: string
      - description: updates
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateServerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Server'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - TelegramInitData: []
      summary: Update server
      tags:
      - admin
  /api/admin/servers/{server_id}/test:
    post:
      parameters:
      - description: server uuid
        in: path
        name: server_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Test server XUI connection
      tags:
      - admin
...
  /api/servers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: List active servers
      tags:
      - servers
  /api/servers/incidents:
    get:
      parameters:
      - default: 168
        description: window in hours
        in: query
        name: hours
        type: integer
      - default: 50
        description: max items
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - TelegramInitData: []
      summary: Recent server incidents
      tags:
      - servers
...
schemes:
- https
securityDefinitions:
  TelegramInitData:
    description: |-
      Raw Telegram.WebApp.initData query string. HMAC-SHA256
      signed by the bot token; freshness window is 1 hour.
    in: header
    name: X-Telegram-Init-Data
    type: apiKey
swagger: "2.0"